pyproject.toml
[project]
name = "polaris"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ai-edge-litert>=2.0.3",
    "numpy==2.2.1",
    "opencv-contrib-python==4.12.0.88",
    "opencv-python==4.12.0.88",
    "pillow==11.1.0",
    "pyntcore==2025.3.2.3",
    "pypylon==4.1.0",
    "robotpy-apriltag>=2025.3.2.3",
    "robotpy-wpimath==2025.3.2.3",
]

[tool.uv]
environments = [
    "sys_platform == 'darwin'",
    "sys_platform == 'linux'",
]
